Move splash into order folder.
