+++ Modules/Wider.Splash/ViewModels/SplashViewModel.cs
@@ -16,7 +16,7 @@ using System;
 using Wider.Interfaces;
 using Wider.Interfaces.Events;
 
-namespace Wider.Splash
+namespace Wider.Splash.ViewModels
 {
     public class SplashViewModel : BindableBase
     {
@@ -38,11 +38,7 @@ namespace Wider.Splash
         public String Status
         {
             get => _status;
-            set
-            {
-                _status = value;
-                RaisePropertyChanged("Status");
-            }
+            set => SetProperty(ref _status, value);
         }
 
         #endregion
