+++ Modules/Wider.Splash/Views/SplashView.xaml
@@ -1,11 +1,20 @@
-﻿<Window x:Class="Wider.Splash.SplashView" xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" x:ClassModifier="internal"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008" xmlns:Behaviours="clr-namespace:Wider.Splash.Behaviours"
-        xmlns:Controls="http://metro.mahapps.com/winfx/xaml/controls" mc:Ignorable="d" d:DesignHeight="279" d:DesignWidth="450"
-        Behaviours:SplashBehaviour.Enabled="True" WindowStartupLocation="CenterScreen" WindowStyle="None" ShowInTaskbar="False"
-        Topmost="True" AllowsTransparency="True" SizeToContent="Manual" Width="540" Height="360" ResizeMode="NoResize"
-        SnapsToDevicePixels="True" Background="Transparent" OpacityMask="Black" Foreground="White">
+﻿<Window 
+    x:Class="Wider.Splash.Views.SplashView" 
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" x:ClassModifier="internal"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+    
+    xmlns:Behaviours="clr-namespace:Wider.Splash.Behaviours"
+    xmlns:Controls="http://metro.mahapps.com/winfx/xaml/controls" 
+    mc:Ignorable="d" d:DesignHeight="279" d:DesignWidth="450"
+        
+    Behaviours:SplashBehaviour.Enabled="True" 
+        
+    WindowStartupLocation="CenterScreen" WindowStyle="None" ShowInTaskbar="False"
+    Topmost="True" AllowsTransparency="True" SizeToContent="Manual" 
+    Width="540" Height="360" ResizeMode="NoResize"
+    SnapsToDevicePixels="True" Background="Transparent" OpacityMask="Black" Foreground="White">
     <Window.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
